+++ ichnaea/api/views.py
@@ -89,7 +89,7 @@ class BaseAPIView(BaseView):
                 skip_check = True
                 self.raven_client.captureException()
 
-        if api_key is not None and api_key.should_allow(self.view_type):
+        if api_key is not None and api_key.allowed(self.view_type):
             self.log_count(api_key.valid_key,
                            api_key.should_log(self.view_type))