+++ ichnaea/tests/factories.py
@@ -96,6 +96,12 @@ class ApiKeyFactory(BaseSQLFactory):
     allow_fallback = False
     allow_locate = True
 
+    fallback_name = 'test_fallback'
+    fallback_url = 'http://127.0.0.1:9/?api'
+    fallback_ratelimit = 10
+    fallback_ratelimit_interval = 60
+    fallback_cache_expire = 60
+
 
 class BboxFactory(Factory):